Refactor dock badge logic into a separate module; Added centralized logging with env_logger
