+++ src-tauri/Cargo.toml
@@ -27,6 +27,5 @@ tauri-plugin-http = "2"
 # You need the tray plugin separately now
 tauri-plugin-shell = "2"
 tauri-plugin-notification = "2"
-objc = "0.2"
-cocoa = "0.26"
-core-foundation = "0.9"
+log = "0.4.26"
+env_logger = "0.11.6"
