+++ src-tauri/src/main.rs
@@ -1,6 +1,8 @@
 // Prevents additional console window on Windows in release, DO NOT REMOVE!!
 #![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
 
+use bitbucket_notifier_lib::tauri_dock_badge;
+
 fn main() {
-    bitbucket_notifier_lib::run()
+    tauri_dock_badge::run()
 }
